refactor: Remove deprecated indexing package and consolidate index management
- Delete obsolete indexing package files and modules
- Move Elasticsearch index management to storage package
- Remove unnecessary indexing-related dependencies from search command
- Simplify index management implementation with direct Elasticsearch client
- Update module configuration to use new index management approach
