+++ internal/storage/index_manager.go
@@ -1,0 +1,8 @@
+package storage
+
+import (
+	"github.com/jonesrussell/gocrawl/internal/api"
+)
+
+// IndexManager is an alias for api.IndexManager to maintain compatibility
+type IndexManager = api.IndexManager
